backend/infra/observability/requirements.txt
# EPIC-OBS-001 Phase 2: Observability Module Dependencies
# Add these to main requirements.txt or install separately

# Structured JSON logging
python-json-logger>=2.0.7

# OpenTelemetry (distributed tracing)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-otlp>=1.21.0

# Note: prometheus-client already exists in requirements.txt
